refactor: extract OrderSummaryDetails component to avoid duplication
